+++ packages/web-components/src/components/global-header/components/global-header/src/components/CommonHeader/__stories__/CommonHeader.stories.ts
@@ -454,7 +454,7 @@ const sidekickConfig = {
   context: 'application',
   insights_enabled: true,
   chat_enabled: true,
-  reports_enabled: true,
+  overview_enabled: true,
   tell_me_more_enabled: true,
 };