+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/HybridIpaasHeader.ts
@@ -46,6 +46,7 @@ export class HybridIpaasHeader extends LitElement {
   @property({ type: Boolean }) solisSidekickEnabled = false;
   @property({ type: Boolean }) solisSwitcherEnabled = false;
   @property({ type: String }) solisEnvironment = 'local';
+  @property({ type: Boolean }) solisDevMode = false; // override for storybook and local to use mock data
   @property({ type: String }) basePath = '';
   @property({ type: String }) displayName = '';
   @property({ type: String }) userEmail = '';
@@ -168,16 +169,16 @@ export class HybridIpaasHeader extends LitElement {
       isEnabled: this.solisSidekickEnabled,
       scriptUrl: SOLIS_CDN_HOSTNAMES[env] + '/sidekick/solis-sidekick.es.js',
       insights_enabled: true,
-      reports_enabled: true,
+      overview_enabled: true,
       chat_enabled: false,
       tell_me_more_enabled: false,
     };
   }
-  private initSolisOptions(env = 'local', forSidekick = false) {
+  private initSolisOptions(env = 'local', forSidekick = false, isProd = true) {
     if (forSidekick) {
       return {
         isEnabled: false,
-        is_prod: env === 'prod',
+        is_prod: isProd,
         cdn_hostname: SOLIS_CDN_HOSTNAMES[env],
         deployment_environment: solisDeploymentEnvironment[env] || 'local',
         product_id: 'ipaas',
@@ -186,7 +187,7 @@ export class HybridIpaasHeader extends LitElement {
       return {
         isEnabled: true,
         scriptUrl: SOLIS_CDN_HOSTNAMES[env] + '/switcher/solis-switcher.es.js',
-        is_prod: env === 'prod',
+        is_prod: isProd,
         cdn_hostname: SOLIS_CDN_HOSTNAMES[env],
         deployment_environment: solisDeploymentEnvironment[env] || 'local',
         product_id: 'ipaas',
@@ -329,14 +330,19 @@ export class HybridIpaasHeader extends LitElement {
     if (this.solisSidekickEnabled) {
       updatedOptions.solisConfig = this.initSolisOptions(
         this.solisEnvironment,
-        true
+        true,
+        !this.solisDevMode
       );
       updatedOptions.sidekickConfig = this.initSidekickOptions(
         this.solisEnvironment
       );
     }
     if (this.solisSwitcherEnabled) {
-      updatedOptions.solisConfig = this.initSolisOptions(this.solisEnvironment);
+      updatedOptions.solisConfig = this.initSolisOptions(
+        this.solisEnvironment,
+        false,
+        !this.solisDevMode
+      );
     }
 
     if (