+++ packages/web-components/src/components/network-graph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.24.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-network-graph@0.23.0...@carbon-labs/wc-network-graph@0.24.0) (2026-01-14)
+
+**Note:** Version bump only for package @carbon-labs/wc-network-graph
+
+
+
+
+
 # [0.23.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-network-graph@0.22.0...@carbon-labs/wc-network-graph@0.23.0) (2025-12-16)